Move `is_inside_unsafe` to `Semantics` impl
